+++ source/stores/decayed-tabs.ts
@@ -7,11 +7,10 @@ import {
   KEEP_N_DECAYED_TABS,
   UPDATE_EVENT_TYPES,
 } from "../lib/constants";
-
-type DecayedTab = browser.Tabs.Tab & { _id: number; id: string };
+import { BrowserTab } from "../lib/types";
 
 const decayedTabs = () => {
-  const { subscribe, update, sync } = cacheable<DecayedTab[]>(
+  const { subscribe, update, sync } = cacheable<BrowserTab[]>(
     DECAY_LOG_CACHE_KEY,
     UPDATE_EVENT_TYPES
   );
@@ -19,18 +18,18 @@ const decayedTabs = () => {
   const add = (decayingTab: browser.Tabs.Tab) => {
     update(value =>
       [
-        { ...decayingTab, _id: decayingTab.id, id: `${DECAYED_TAB_PREFIX}-{temp}` } as DecayedTab,
+        { ...decayingTab, _id: decayingTab.id, id: `${DECAYED_TAB_PREFIX}-{temp}` } as BrowserTab,
         ...(value
           ? value.filter(decayedTab => decayingTab.url !== decayedTab.url)
-          : ([] as DecayedTab[])),
+          : ([] as BrowserTab[])),
       ]
         .slice(0, KEEP_N_DECAYED_TABS)
         .map(
-          (tab: DecayedTab, idx: number) =>
+          (tab: BrowserTab, idx: number) =>
             ({
               ...tab,
               id: `${DECAYED_TAB_PREFIX}-${idx}`,
-            }) as DecayedTab
+            }) as BrowserTab
         )
     );
   };
